refactor(zuuli): isolate native send policy authority
