+++ wallet/plugins/tauri-plugin-zcash/src/wallet/send.rs
@@ -11,23 +11,12 @@ use std::{
 use rand::{RngCore, rngs::OsRng};
 use secrecy::ExposeSecret;
 use sha2::{Digest, Sha256};
-use zcash_client_backend::data_api::wallet::{
-    ConfirmationsPolicy, CreateErrT, LockRequest, ProposeTransferErrT, SpendingKeys,
-    create_proposed_transactions,
-    input_selection::{GreedyInputSelector, SpendPolicy},
-    propose_transfer,
-};
-use zcash_client_backend::data_api::{InputSource, WalletCommitmentTrees, WalletRead, WalletWrite};
-use zcash_client_backend::fees::DustOutputPolicy;
-use zcash_client_backend::fees::zip317::SingleOutputChangeStrategy;
-use zcash_client_backend::proposal::Proposal;
+use zcash_client_backend::data_api::WalletRead;
+use zcash_client_backend::data_api::wallet::{ConfirmationsPolicy, SpendingKeys};
 use zcash_client_backend::proto::service::{RawTransaction, TxFilter};
-use zcash_client_backend::wallet::OvkPolicy;
 use zcash_keys::address::Address;
-use zcash_primitives::transaction::{TxVersion, fees::zip317::FeeRule as Zip317FeeRule};
 use zcash_proofs::prover::LocalTxProver;
 use zcash_protocol::PoolType;
-use zcash_protocol::consensus;
 use zcash_protocol::memo::{Memo, MemoBytes};
 use zcash_protocol::value::Zatoshis;
 use zcash_protocol::{ShieldedPool, TxId};
@@ -40,6 +29,10 @@ use crate::models::{
 use crate::wallet::client::connect_to_lightwalletd;
 use crate::wallet::keys;
 use crate::wallet::{WalletProposal, WalletState};
+
+mod native;
+
+use native::{create_transactions, propose_fixed, propose_send_all_attempt};
 
 /// A transaction that has already been created in the wallet database.
 /// Retrying this record always rebroadcasts `raw_transaction`; it never signs
@@ -74,156 +67,6 @@ const SEND_CONFIRMATION_TOKEN_DOMAIN: &[u8] = b"ZUULI_SEND_CONFIRMATION_TOKEN\0"
 const SEND_FEE_POLICY: &str = "zip317-standard";
 const SEND_CHANGE_POLICY: &str = "zip317-shielded-auto";
 const SEND_CONFIRMATION_TTL: Duration = Duration::from_secs(2 * 60);
-
-fn proposal_lock_request() -> Option<LockRequest> {
-    // The process-local send state machine already serializes proposal and
-    // execution. Preserve the pre-locking API behavior until durable owner and
-    // unlock recovery semantics are designed together.
-    None
-}
-
-fn proposed_transaction_version() -> Option<TxVersion> {
-    // Let librustzcash select the consensus transaction version for the target
-    // height, matching the behavior before this argument was introduced.
-    None
-}
-
-type NativeSendChangeStrategy<DbT> = SingleOutputChangeStrategy<Zip317FeeRule, DbT>;
-type NativeSendProposalError<DbT> = ProposeTransferErrT<
-    DbT,
-    zcash_client_sqlite::wallet::commitment_tree::Error,
-    GreedyInputSelector<DbT>,
-    NativeSendChangeStrategy<DbT>,
->;
-
-/// The single production authority for money-affecting proposal policy. Both
-/// fixed-amount sends and every send-all retry cross this exact boundary.
-#[allow(clippy::type_complexity)]
-fn propose_native_send_with_policy<DbT, ParamsT>(
-    db: &mut DbT,
-    params: &ParamsT,
-    account_id: <DbT as InputSource>::AccountId,
-    request: zip321::TransactionRequest,
-    spend_policy: &SpendPolicy,
-) -> std::result::Result<
-    Proposal<Zip317FeeRule, <DbT as InputSource>::NoteRef>,
-    NativeSendProposalError<DbT>,
->
-where
-    DbT: WalletWrite + InputSource<Error = <DbT as WalletRead>::Error>,
-    <DbT as InputSource>::NoteRef: Copy + Eq + Ord,
-    ParamsT: consensus::Parameters + Clone,
-{
-    let input_selector = GreedyInputSelector::new();
-    let change_strategy = SingleOutputChangeStrategy::new(
-        Zip317FeeRule::standard(),
-        None,
-        ShieldedPool::Orchard,
-        DustOutputPolicy::default(),
-    );
-
-    propose_transfer::<_, _, _, _, zcash_client_sqlite::wallet::commitment_tree::Error>(
-        db,
-        params,
-        account_id,
-        &input_selector,
-        &change_strategy,
-        request,
-        ConfirmationsPolicy::default(),
-        spend_policy,
-        proposal_lock_request(),
-        proposed_transaction_version(),
-    )
-}
-
-#[allow(clippy::type_complexity)]
-fn propose_native_send<DbT, ParamsT>(
-    db: &mut DbT,
-    params: &ParamsT,
-    account_id: <DbT as InputSource>::AccountId,
-    request: zip321::TransactionRequest,
-) -> std::result::Result<
-    Proposal<Zip317FeeRule, <DbT as InputSource>::NoteRef>,
-    NativeSendProposalError<DbT>,
->
-where
-    DbT: WalletWrite + InputSource<Error = <DbT as WalletRead>::Error>,
-    <DbT as InputSource>::NoteRef: Copy + Eq + Ord,
-    ParamsT: consensus::Parameters + Clone,
-{
-    propose_native_send_with_policy(db, params, account_id, request, &SpendPolicy::default())
-}
-
-#[allow(clippy::type_complexity)]
-fn propose_fixed_native_send<DbT, ParamsT>(
-    db: &mut DbT,
-    params: &ParamsT,
-    account_id: <DbT as InputSource>::AccountId,
-    request: zip321::TransactionRequest,
-) -> std::result::Result<
-    Proposal<Zip317FeeRule, <DbT as InputSource>::NoteRef>,
-    NativeSendProposalError<DbT>,
->
-where
-    DbT: WalletWrite + InputSource<Error = <DbT as WalletRead>::Error>,
-    <DbT as InputSource>::NoteRef: Copy + Eq + Ord,
-    ParamsT: consensus::Parameters + Clone,
-{
-    propose_native_send(db, params, account_id, request)
-}
-
-#[allow(clippy::type_complexity)]
-fn propose_send_all_native_attempt<DbT, ParamsT>(
-    db: &mut DbT,
-    params: &ParamsT,
-    account_id: <DbT as InputSource>::AccountId,
-    request: zip321::TransactionRequest,
-) -> std::result::Result<
-    Proposal<Zip317FeeRule, <DbT as InputSource>::NoteRef>,
-    NativeSendProposalError<DbT>,
->
-where
-    DbT: WalletWrite + InputSource<Error = <DbT as WalletRead>::Error>,
-    <DbT as InputSource>::NoteRef: Copy + Eq + Ord,
-    ParamsT: consensus::Parameters + Clone,
-{
-    propose_native_send(db, params, account_id, request)
-}
-
-type NativeSendCreationError<DbT> = CreateErrT<
-    DbT,
-    std::convert::Infallible,
-    Zip317FeeRule,
-    std::convert::Infallible,
-    <DbT as InputSource>::NoteRef,
->;
-
-/// The single production authority for transaction creation and outgoing
-/// viewing-key retention.
-#[allow(clippy::type_complexity)]
-fn create_native_send_transactions<DbT, ParamsT>(
-    db: &mut DbT,
-    params: &ParamsT,
-    prover: &LocalTxProver,
-    spending_keys: &SpendingKeys,
-    proposal: &Proposal<Zip317FeeRule, <DbT as InputSource>::NoteRef>,
-) -> std::result::Result<nonempty::NonEmpty<TxId>, NativeSendCreationError<DbT>>
-where
-    DbT: WalletWrite + WalletCommitmentTrees + InputSource,
-    ParamsT: consensus::Parameters + Clone,
-{
-    create_proposed_transactions::<_, _, std::convert::Infallible, _, std::convert::Infallible, _>(
-        db,
-        params,
-        prover,
-        prover,
-        spending_keys,
-        OvkPolicy::Sender,
-        proposal,
-        // Preserve the builder-derived expiry height for every proposal step.
-        None,
-    )
-}
 
 #[derive(Clone, Copy)]
 struct ConfirmationClock {
@@ -1543,7 +1386,7 @@ pub async fn propose_send(
         .copied()
         .ok_or(Error::SendError("no accounts found".into()))?;
 
-    let proposal = propose_fixed_native_send(db, &state.network, account_id, request)
+    let proposal = propose_fixed(db, &state.network, account_id, request)
         .map_err(|e| Error::SendError(format!("failed to propose transfer: {e:?}")));
 
     drop(db_guard);
@@ -1679,7 +1522,7 @@ pub async fn propose_send_all(
             .copied()
             .ok_or(Error::SendError("no accounts found".into()))?;
 
-        let result = propose_send_all_native_attempt(db, &state.network, account_id, request);
+        let result = propose_send_all_attempt(db, &state.network, account_id, request);
 
         drop(db_guard);
 
@@ -1905,7 +1748,7 @@ pub async fn execute_send(
 
     let mut transaction_created = false;
     let created = (|| -> Result<PendingBroadcast> {
-        let txids = create_native_send_transactions(
+        let txids = create_transactions(
             db,
             &state.network,
             prover,
@@ -1915,7 +1758,7 @@ pub async fn execute_send(
         .map_err(|e| Error::SendError(format!("failed to create transaction: {e:?}")))?;
         transaction_created = true;
 
-        // `create_proposed_transactions` returns `NonEmpty<TxId>`.
+        // The native creation boundary returns `NonEmpty<TxId>`.
         let txid = *txids.first();
         let tx = db
             .get_transaction(txid)
@@ -2369,19 +2212,7 @@ async fn broadcast_record(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use transparent::{
-        bundle::{OutPoint, TxOut},
-        keys::TransparentKeyScope,
-    };
     use zcash_address::unified::{Address as UnifiedAddress, Encoding, Receiver};
-    use zcash_client_backend::data_api::Account;
-    use zcash_client_backend::data_api::testing::{
-        orchard::OrchardPoolTester,
-        pool::{ShieldedPoolTester, dsl::TestDsl},
-    };
-    use zcash_client_backend::data_api::wallet::input_selection::TransparentSpendPolicy;
-    use zcash_client_backend::wallet::WalletTransparentOutput;
-    use zcash_client_sqlite::testing::{BlockCache, db::TestDbFactory};
     use zcash_protocol::consensus::{Network, NetworkType};
 
     const MAINNET_TADDR: &str = "t1Hsc1LR8yKnbbe3twRp88p6vFfC5t7DLbs";
@@ -2392,201 +2223,6 @@ mod tests {
     const WALLET_ID: &str = "wallet-a";
     const SESSION_ID: [u8; 32] = [0x11; 32];
     const OTHER_SESSION_ID: [u8; 32] = [0x22; 32];
-
-    fn payment_request(
-        network: &impl consensus::Parameters,
-        recipient: &Address,
-        amount: u64,
-    ) -> zip321::TransactionRequest {
-        zip321::TransactionRequest::new(vec![zip321::Payment::without_memo(
-            recipient.to_zcash_address(network),
-            Zatoshis::const_from_u64(amount),
-        )])
-        .expect("the independently fixed test payment is valid")
-    }
-
-    #[test]
-    fn production_send_boundaries_preserve_pool_depth_fee_and_sender_recovery() {
-        const NOTE_VALUE: u64 = 60_000;
-        const FIXED_SEND_VALUE: u64 = 10_000;
-        const SEND_ALL_VALUE: u64 = 50_000;
-        const EXPECTED_FEE: u64 = 10_000;
-
-        let mut st =
-            TestDsl::with_sapling_birthday_account(TestDbFactory::default(), BlockCache::new())
-                .build::<OrchardPoolTester>();
-        let (_, _, _) = st.add_a_single_note_checking_balance(Zatoshis::const_from_u64(NOTE_VALUE));
-
-        let recipient_key = OrchardPoolTester::sk(&[0xf5; 32]);
-        let recipient = OrchardPoolTester::sk_default_address(&recipient_key);
-        let account = st.get_account();
-        let account_id = account.id();
-        let network = *st.network();
-
-        let immature = propose_fixed_native_send(
-            st.wallet_mut(),
-            &network,
-            account_id,
-            payment_request(&network, &recipient, FIXED_SEND_VALUE),
-        );
-        assert!(
-            immature.is_err(),
-            "an externally received Orchard note must not be spendable after one confirmation"
-        );
-
-        st.add_empty_blocks(9);
-        let proposal = propose_fixed_native_send(
-            st.wallet_mut(),
-            &network,
-            account_id,
-            payment_request(&network, &recipient, FIXED_SEND_VALUE),
-        )
-        .expect("the same Orchard note must be spendable after ten confirmations");
-        assert_eq!(proposal.steps().len(), 1);
-        assert_eq!(proposal.input_count_in_pool(PoolType::ORCHARD), 1);
-        assert_eq!(proposal.input_count_in_pool(PoolType::SAPLING), 0);
-        assert_eq!(
-            proposal
-                .steps()
-                .head
-                .change_count_in_pool(PoolType::ORCHARD),
-            1
-        );
-        assert_eq!(
-            proposal
-                .steps()
-                .head
-                .change_count_in_pool(PoolType::SAPLING),
-            0
-        );
-        assert_eq!(
-            u64::from(proposal.steps().head.balance().fee_required()),
-            EXPECTED_FEE
-        );
-
-        // Exercise the production wrapper used by every send-all retry. Spending the
-        // independently fixed note value minus the expected fee must consume the whole note,
-        // retain ZIP 317's fee, and return no value as change. Orchard padding can retain a
-        // zero-valued change entry, so value rather than vector shape is the behavior at stake.
-        let send_all_proposal = propose_send_all_native_attempt(
-            st.wallet_mut(),
-            &network,
-            account_id,
-            payment_request(&network, &recipient, SEND_ALL_VALUE),
-        )
-        .expect("the send-all-shaped request must be proposed by the shared boundary");
-        assert_eq!(send_all_proposal.input_count_in_pool(PoolType::ORCHARD), 1);
-        assert_eq!(
-            u64::from(send_all_proposal.steps().head.balance().fee_required()),
-            EXPECTED_FEE
-        );
-        assert_eq!(
-            send_all_proposal
-                .steps()
-                .head
-                .balance()
-                .proposed_change()
-                .iter()
-                .map(|change| u64::from(change.value()))
-                .sum::<u64>(),
-            0
-        );
-
-        let spending_keys = SpendingKeys::from_unified_spending_key(account.usk().clone());
-        let prover = LocalTxProver::bundled();
-        let recovery_height = proposal.min_target_height().into();
-        let txids = create_native_send_transactions(
-            st.wallet_mut(),
-            &network,
-            &prover,
-            &spending_keys,
-            &proposal,
-        )
-        .expect("the production creation boundary must build and persist the transaction");
-        let tx = st
-            .wallet()
-            .get_transaction(txids.head)
-            .expect("the test wallet database remains readable")
-            .expect("the created transaction is persisted");
-        let sender_fvk = OrchardPoolTester::test_account_fvk(&st);
-        let (_, recovered_recipient, _) =
-            OrchardPoolTester::try_output_recovery(&network, recovery_height, &tx, &sender_fvk)
-                .expect("the sender OVK must recover the external Orchard output");
-        assert_eq!(recovered_recipient, recipient);
-    }
-
-    #[test]
-    fn production_transparent_spend_shields_change_to_orchard() {
-        const UTXO_VALUE: u64 = 60_000;
-        const PAYMENT_VALUE: u64 = 10_000;
-        const EXPECTED_CHANGE: u64 = 35_000;
-
-        let mut st =
-            TestDsl::with_sapling_birthday_account(TestDbFactory::default(), BlockCache::new())
-                .build::<OrchardPoolTester>();
-        let account = st.get_account();
-        let account_id = account.id();
-        let network = *st.network();
-        let (transparent_recipient, _) = account.usk().default_transparent_address();
-        let received_height = st.add_empty_blocks(1);
-        let utxo = WalletTransparentOutput::from_parts(
-            OutPoint::fake(),
-            TxOut::new(
-                Zatoshis::const_from_u64(UTXO_VALUE),
-                transparent_recipient.script().into(),
-            ),
-            Some(received_height),
-            Some(account_id),
-            Some(TransparentKeyScope::EXTERNAL),
-            None,
-        )
-        .expect("the independently constructed transparent UTXO is valid");
-        st.wallet_mut()
-            .put_received_transparent_utxo(&utxo)
-            .expect("the real wallet database accepts the transparent UTXO");
-        st.add_empty_blocks(9);
-
-        let spend_policy =
-            SpendPolicy::default().with_transparent(TransparentSpendPolicy::any_account_addr());
-        let proposal = propose_native_send_with_policy(
-            st.wallet_mut(),
-            &network,
-            account_id,
-            payment_request(
-                &network,
-                &Address::from(transparent_recipient),
-                PAYMENT_VALUE,
-            ),
-            &spend_policy,
-        )
-        .expect("the production proposal core proposes the mature transparent spend");
-        assert_eq!(proposal.input_count_in_pool(PoolType::TRANSPARENT), 1);
-        assert_eq!(
-            proposal
-                .steps()
-                .head
-                .change_count_in_pool(PoolType::ORCHARD),
-            1
-        );
-        assert_eq!(
-            proposal
-                .steps()
-                .head
-                .change_count_in_pool(PoolType::SAPLING),
-            0
-        );
-        assert_eq!(
-            proposal
-                .steps()
-                .head
-                .balance()
-                .proposed_change()
-                .iter()
-                .map(|change| u64::from(change.value()))
-                .sum::<u64>(),
-            EXPECTED_CHANGE
-        );
-    }
 
     fn pending(status: BroadcastStatus) -> PendingBroadcast {
         PendingBroadcast {
